refactor of class placement
